# api/openapi.yaml
openapi: 3.1.0
info:
  title: Collections Agent API
  version: 1.0.0
  description: REST API for the AI Receivables Collections Agent console, the bank webhook and admin controls. Serves the
    stories named in each operation's x-story-ids. Money fields are integer paise named *_paise (INR only); every amount shown
    to a user is formatted from these at the edge.
  contact:
    name: HACK-001 team
...
servers:
- url: http://localhost:8000/api/v1
  description: local Docker Compose
security:
- demoRole: []
...
paths:
...
  /priorities:
    get:
      operationId: listPriorities
      summary: Top customers by priority with reasons (top 15 by default)
      tags:
      - customers
      x-story-ids:
      - US-00-002
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Priority'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers:
    get:
      operationId: listCustomers
      summary: CRM list
      tags:
      - customers
      x-story-ids:
      - US-00-026
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: filter[band]
        in: query
        required: false
        schema:
          type: string
          enum:
          - HIGH
          - MEDIUM
          - LOW
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers/{id}:
    get:
      operationId: getCustomer
      summary: Customer detail header
      tags:
      - customers
      x-story-ids:
      - US-00-004
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers/{id}/invoices:
    get:
      operationId: listCustomerInvoices
      summary: Invoices with derived balances
      tags:
      - customers
      x-story-ids:
      - US-00-001
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: filter[status]
        in: query
        required: false
        schema:
          type: string
          enum:
          - unpaid
          - partially_paid
          - paid
          - disputed
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - page
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Invoice'
                  page:
                    $ref: '#/components/schemas/Page'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /customers/{id}/priority:
    get:
      operationId: getCustomerPriority
      summary: Score, band and reasons
      tags:
      - customers
      x-story-ids:
      - US-00-002
      - US-00-004
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Priority'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /calls:
    post:
      operationId: startCall
      summary: Start an AI voice call (voice flag, kill switch, consent, undisputed balance)
      tags:
      - calls
      x-story-ids:
      - HACK-003
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - customer_id
              properties:
                customer_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
    get:
      operationId: listCalls
      summary: Calls with turns, outcome and summary
      tags:
      - calls
      x-story-ids:
      - HACK-003
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: filter[customer_id]
        in: query
        required: false
        schema:
          type: string
      - name: limit
        in: query
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /calls/{id}:
    get:
      operationId: getCall
      summary: One call with its turns
      tags:
      - calls
      x-story-ids:
      - HACK-003
      x-roles: viewer, collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /calls/{id}/turns:
    post:
      operationId: customerSays
      summary: 'SIMULATED calls only: the customer''s words, typed by a collector'
      tags:
      - calls
      x-story-ids:
      - HACK-003
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required:
              - text
              properties:
                text:
                  type: string
                  minLength: 1
                  maxLength: 2000
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
        '500':
          $ref: '#/components/responses/InternalError'
  /calls/{id}/end:
    post:
      operationId: endCall
      summary: End a call; the summary is written from the call row
      tags:
      - calls
      x-story-ids:
      - HACK-003
      x-roles: collector, admin
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /webhooks/voice/{id}/turn:
    post:
      operationId: voiceTurnWebhook
      summary: Twilio speech result; answers TwiML. X-Twilio-Signature required
      tags:
      - calls
      x-story-ids:
      - HACK-003
      x-roles: public (signed)
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /webhooks/voice/{id}/status:
    post:
      operationId: voiceStatusWebhook
      summary: Twilio call status callback. X-Twilio-Signature required
      tags:
      - calls
      x-story-ids:
      - HACK-003
      x-roles: public (signed)
      parameters:
      - $ref: '#/components/parameters/ApiVersion'
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          headers:
            X-Request-Id:
              $ref: '#/components/headers/XRequestId'
          content:
            application/json:
              schema:
                type: object
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    demoRole:
      type: apiKey
      in: header
      name: X-Demo-Role
      description: admin, collector or viewer
    bankSignature:
      type: apiKey
      in: header
      name: X-Bank-Signature
      description: hex HMAC-SHA256 of '<X-Bank-Timestamp>.<raw body>' with BANK_WEBHOOK_SECRET
  parameters:
    ApiVersion:
      name: X-API-Version
      in: header
      required: false
      schema:
        type: string
        example: '2026-09-30'
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  headers:
    XRequestId:
      schema:
        type: string
      description: Matches error.request_id and the server log line
  responses:
...
    Unauthorized:
      description: Not signed in
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: UNAUTHORIZED
              message: Not signed in
              request_id: req_01J9Z6
    Forbidden:
      description: Signed in but not allowed
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: FORBIDDEN
              message: Signed in but not allowed
              request_id: req_01J9Z6
    NotFound:
      description: Not found
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: NOT_FOUND
              message: Not found
              request_id: req_01J9Z6
...
    UnprocessableEntity:
      description: Valid shape, invalid content
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: AMOUNT_MISMATCH
              message: Valid shape, invalid content
              request_id: req_01J9Z6
...
    InternalError:
      description: Unexpected error
      headers:
        X-Request-Id:
          $ref: '#/components/headers/XRequestId'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: INTERNAL
              message: Unexpected error
              request_id: req_01J9Z6
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Stable UPPER_SNAKE code, e.g. AMOUNT_MISMATCH, NOT_APPROVED, STALE_DRAFT
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  reason:
                    type: string
                required:
                - field
                - reason
            request_id:
              type: string
          required:
          - code
          - message
          - request_id
      required:
      - error
      example:
        error:
          code: AMOUNT_MISMATCH
          message: ₹4,50,000 is not the ledger amount for INV-1021
          details:
          - field: body
            reason: INV-1021 is 40000000 paise
          request_id: req_01J9Z6
    Page:
      type: object
      properties:
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
        has_more:
          type: boolean
      required:
      - next_cursor
      - has_more
      example:
        next_cursor: eyJpZCI6IjAxOTIifQ
        has_more: false
...
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        segment:
          type: string
          enum:
          - enterprise
          - mid_market
          - sme
        credit_terms_days:
          type: integer
        outstanding_paise:
          type: integer
          format: int64
          minimum: 0
          description: Total outstanding (integer paise, INR)
        overdue_paise:
          type: integer
          format: int64
          minimum: 0
          description: Total overdue (integer paise, INR)
        band:
          anyOf:
          - type: string
            enum:
            - HIGH
            - MEDIUM
            - LOW
          - type: 'null'
        last_contact_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
        next_action:
          anyOf:
          - type: string
          - type: 'null'
      required:
      - id
      - name
      - email
      - phone
      - segment
      - credit_terms_days
      - outstanding_paise
      - overdue_paise
      - band
      - last_contact_at
      - next_action
      example:
        id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        name: ABC Distributors
        email: accounts@abc-distributors.example.in
        phone: +91-9800000001
        segment: mid_market
        credit_terms_days: 30
        outstanding_paise: 75000000
        overdue_paise: 75000000
        band: HIGH
        last_contact_at: null
        next_action: Send follow-up on missed promise
    Invoice:
      type: object
      properties:
        id:
          type: string
          format: uuid
        number:
          type: string
          pattern: ^INV-[0-9]+$
        invoice_date:
          type: string
          format: date
        due_date:
          type: string
          format: date
        amount_paise:
          type: integer
          format: int64
          minimum: 0
          description: Invoice amount (integer paise, INR)
        paid_paise:
          type: integer
          format: int64
          minimum: 0
          description: Sum of allocations (integer paise, INR)
        remaining_paise:
          type: integer
          format: int64
          minimum: 0
          description: Amount minus paid (integer paise, INR)
        status:
          type: string
          enum:
          - unpaid
          - partially_paid
          - paid
          - disputed
        days_overdue:
          type: integer
          description: From the demo clock; 0 when not due
      required:
      - id
      - number
      - invoice_date
      - due_date
      - amount_paise
      - paid_paise
      - remaining_paise
      - status
      - days_overdue
      example:
        id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        number: INV-1021
        invoice_date: '2026-08-12'
        due_date: '2026-09-11'
        amount_paise: 40000000
        paid_paise: 0
        remaining_paise: 40000000
        status: unpaid
        days_overdue: 19
    Priority:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        customer_name:
          type: string
        score:
          type: integer
          minimum: 0
          maximum: 100
        band:
          type: string
          enum:
          - HIGH
          - MEDIUM
          - LOW
        reasons:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              text:
                type: string
            required:
            - code
            - text
        factors:
          type: array
          description: Each scored input with the points it added (decision factors, not model reasoning).
          items:
            type: object
            properties:
              code:
                type: string
              label:
                type: string
              value:
                type: string
              points:
                type: number
              rule:
                type: string
            required:
            - code
            - label
            - value
            - points
            - rule
      required:
      - customer_id
      - customer_name
      - score
      - band
      - reasons
      - factors
      example:
        customer_id: 0192b1c2-7f3a-7c4e-9a1b-2c3d4e5f6a7b
        customer_name: ABC Distributors
        score: 86
        band: HIGH
        reasons:
        - code: HIGH_OUTSTANDING
          text: High outstanding (₹7,50,000)
        - code: OLDEST_OVERDUE
          text: Oldest invoice 19 days overdue
        - code: MISSED_PROMISES
          text: 1 missed promise
...